% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{compare_runs}
\alias{compare_runs}
\title{Create a stacked bar plot of multiple runs}
\usage{
compare_runs(
  csv_folder,
  taxon_level = "species",
  size_col = "Total_reads",
  type = "value"
)
}
\arguments{
\item{taxon_level}{taxonomy level to be plotted. Either family, genus or species}

\item{size_col}{Column from the csv file that is used for the plot}

\item{type}{Default is count to use the values. Change to pct to plot the percentages.}

\item{csv_file}{Path to the input folder}
}
\description{
This function loads multiple ViruScreen output file as a dataframe and
creates a stacked bar plot
}
