% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Create a heatmap}
\usage{
make_heatmap(csv_file)
}
\arguments{
\item{csv_file}{Path to the input file}
}
\description{
This function loads a ViruScreen output file as a dataframe and creates a heatmap
 of all columns
}
