% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{make_nmds}
\alias{make_nmds}
\title{Create a NMDS plot}
\usage{
make_nmds(csv_folder, show_species = FALSE)
}
\arguments{
\item{csv_file}{Path to the input file}
}
\description{
This function loads multiple ViruScreen output file as a dataframe and
creates a NMDS plot
}
