% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{make_sunburst}
\alias{make_sunburst}
\title{Create a sunburst}
\usage{
make_sunburst(csv_file, size_col = "Total_reads")
}
\arguments{
\item{csv_file}{Path to the input file}
}
\description{
This function loads a ViruScreen output file as a dataframe and creates a sunburst of the
taxonomy and total reads
}
