% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{compare_runs}
\alias{compare_runs}
\title{Create a stacked bar plot of multiple runs}
\usage{
compare_runs(
  csv_folder,
  taxon_level = "species",
  size_col = "Total_reads",
  type = "value"
)
}
\arguments{
\item{csv_file}{Path to the input folder}
}
\description{
This function loads multiple ViruScreen output file as a dataframe and
creates a stacked bar plot
}
